+++ docs/backends/analytical_kinematics/files/02_inverse_kinematics.py
@@ -1,7 +1,7 @@
 from compas.geometry import Frame
 
 from compas_fab.backends import AnalyticalKinematicsPlanner
-from compas_fab.backends.kinematics.solvers import UR5Kinematics
+from compas_fab.backends import UR5Kinematics
 from compas_fab.robots import FrameTarget
 from compas_fab.robots import RobotCellLibrary
 from compas_fab.robots import TargetMode