+++ docs/backends/analytical_kinematics/files/04_cartesian_path_analytic_pybullet.py
@@ -1,3 +1,7 @@
+# This example plots the planned trajectory to a window using matplotlib.
+# matplotlib is not a dependency of compas_fab and must be installed separately:
+#   uv pip install matplotlib
+
 import matplotlib.pyplot as plt
 from compas.geometry import Frame